+++ src/infra/http/controllers/create-question.controller.ts
@@ -2,7 +2,7 @@ import { CurrentUser } from '@/infra/auth/current-user-context';
 import { JwtAuthGuard } from '@/infra/auth/jwt-auth.guard';
 import type { UserPayload } from '@/infra/auth/jwt.strategy';
 import { ZodValidationPipe } from '@/infra/http/pipe/zod-validation-pipe';
-import { PrismaService } from '@/infra/prisma/prisma.service';
+import { PrismaService } from '@/infra/database/prisma/prisma.service';
 import { Body, Controller, Post, UseGuards } from '@nestjs/common';
 import z from 'zod';
