+++ src/infra/http/controllers/get-recent-questions.controller.ts
@@ -1,6 +1,6 @@
 import { JwtAuthGuard } from '@/infra/auth/jwt-auth.guard';
 import { ZodValidationPipe } from '@/infra/http/pipe/zod-validation-pipe';
-import { PrismaService } from '@/infra/prisma/prisma.service';
+import { PrismaService } from '@/infra/database/prisma/prisma.service';
 import { Controller, Get, Query, UseGuards } from '@nestjs/common';
 import z from 'zod';
