+++ src/infra/http/http.module.ts
@@ -3,15 +3,16 @@ import { AuthenticateController } from './controllers/authenticate.controller';
 import { CreateAccountController } from './controllers/create-account.controller';
 import { CreateQuestionController } from './controllers/create-question.controller';
 import { GetRecentQuestionsController } from './controllers/get-recent-questions.controller';
-import { PrismaService } from '../prisma/prisma.service';
+import { PrismaService } from '../database/prisma/prisma.service';
+import { DatabaseModule } from '../database/database.module';
 
 @Module({
+  imports: [DatabaseModule],
   controllers: [
     CreateAccountController,
     AuthenticateController,
     CreateQuestionController,
     GetRecentQuestionsController,
   ],
-  providers: [PrismaService],
 })
 export class HttpModule {}
