--- src/infra/prisma/prisma.service.ts
@@ -1,19 +1,0 @@
-import { Injectable, OnModuleDestroy, OnModuleInit } from '@nestjs/common';
-import { PrismaClient } from 'generated/prisma';
-
-@Injectable()
-export class PrismaService extends PrismaClient implements OnModuleInit, OnModuleDestroy {
-  constructor() {
-    super({
-      log: ['warn', 'error'],
-    });
-  }
-
-  onModuleInit() {
-    return this.$connect();
-  }
-
-  onModuleDestroy() {
-    return this.$disconnect();
-  }
-}
